+++ libs/small-steps/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0.0
 
+* Add `withJudgmentContext`
 * Add an `Ord` constraint to the `PredicateFailure` type in the `STS` class
 
 ## 1.1.4.0